fix(nvim-jdtls) Remove unnecessary attach check.
I've removed pointless code believed to check if we were already within
the current project, then not to start up another lsp client. This
shouldn't be the case. nvim-jdtls is built to handle multple projects
open at one time. These lsp processes stay running for the duration of
the neovim session for efficiency purposes. This is by design.