docs(snacks.nvim): Add note about 'sqlite3.dll' file
There's a problem when first using the fuzzy finder 'smart' that's part of the snacks groups of plugins. There's a script that will try and download a required file using powershell in the background, and even though from a safe source, antivirus software hates this. Just download it manually and at it to the standard nvim cache directory.