fix(mini-surround): Mitigate user input issue.
Originally I tried to modify the plugin spec so the user would only have to worry about changing the surrounding under the cursor. This is already achieved in the original
spec, however is a little unclear. I've reverted back to this spec with a few added comments for clarity, as these 'extra' options do prove useful from time to time. Plus
there was an issue with my change where the user wouldn't be prompted again for another surrounding to replace or delete, after first using the keymap. This is because under
the hood the plugin uses neovim operators and caching to remember what the user have previously selected. This isn't good so I'm reverting with a better understanding of the plugin.